Updated controller for ProductosController, ProveedorController& FacturaController along with the update method for all with validation

+++ Ferreteria/src/main/java/com/kodigo/Ferreteria/DAO/ProductosDao.java
@@ -3,5 +3,5 @@ package com.kodigo.Ferreteria.DAO;
 import com.kodigo.Ferreteria.entity.ProductosEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductosDao extends JpaRepository<ProductosEntity, Integer> {
+public interface ProductosDao extends JpaRepository<ProductosEntity, Long> {
 }
